requirements.txt
GoogleAuth
GoogleDrive
auth
GoogleCredentials
gspread
default
pandas
numpy
seaborn
streamlit
opencv-python-headless
scikit_learn
joblib
pillow
pandas
matplotlib
streamlit_drawable_canvas